Move welcome view to new file
